Move these 2 tests from test_users_account.py into test_collections.py -test_user_my_collections_page -test_user_my_favorites_page https://www.pivotaltracker.com/story/show/28492459
